+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- `IriRef::relative_to`
+
 ## [1.2.0] - 2020-09-10
 ### Added
 - A `CHANGELOG.md` file.